lambda_code/requirements.txt
instagrapi==2.0.3
openai==1.11.1
Pillow==10.2.0
Requests==2.31.0
